vehicle_stats/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
motor==3.3.2
pymongo==4.6.0
ultralytics==8.0.227
opencv-python==4.8.1.78
opencv-python-headless==4.8.1.78
Pillow==10.1.0
pandas==2.1.3
numpy==1.26.2
python-dotenv==1.0.0
pydantic==2.5.0
